% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitr-package.R
\docType{package}
\name{bitr-package}
\alias{bitr}
\alias{bitr-package}
\title{Biological Id translatoR}
\description{
converting biological IDs
}
\details{
\tabular{ll}{ Package: \tab bitr\cr Type: \tab Package\cr Version: \tab
0.1\cr Date: \tab 2014-07-28\cr License: \tab Artistic-2.0 \cr LazyLoad:
\tab yes\cr }
}
\author{
Guangchuang Yu

Maintainer: Guangchuang Yu <guangchuangyu@gmail.com>
}
\keyword{package}

