% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EG2ACCNUM.R
\name{EG2ACCNUM}
\alias{EG2ACCNUM}
\title{EGACCNUM}
\usage{
EG2ACCNUM(eg, organism = "human", type = "gene")
}
\arguments{
\item{eg}{Entrez gene ids}

\item{organism}{organism, only human supported yet.}

\item{type}{one of "gene" or "protein"}
}
\value{
a data.frame of eg to accession mapping
}
\description{
Convert Entrez to ACCNUM
}
\details{
Given a character vector of entrez gene ids, this function will convert it to Accession numbers.
}
\examples{
EG2ACCNUM("8061")
}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\seealso{
\code{\link{ACCNUM2EG}}
}
\keyword{manip}

