% Generated by roxygen2 (4.0.1): do not edit by hand
\name{EG2Symbol}
\alias{EG2Symbol}
\title{Convert Entrez to Gene symbol}
\usage{
EG2Symbol(eg, organism = "human", useBiomart = FALSE)
}
\arguments{
\item{eg}{Entrez IDs.}

\item{organism}{organism.}

\item{useBiomart}{using Biomart to query or using annotation packages.}
}
\value{
\item{result}{eg.symbol dataframe.}
}
\description{
Given a character vector of Entrez IDs, this function will convert it to
Gene symbols.
}
\examples{
EG2Symbol("9961")
}
\keyword{manip}

