% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EG2Uniprot.R
\name{EG2Uniprot}
\alias{EG2Uniprot}
\title{Convert Entrez to uniprot ID}
\usage{
EG2Uniprot(eg, organism = "human")
}
\arguments{
\item{eg}{Entrez IDs.}

\item{organism}{organism.}
}
\value{
\item{result}{eg.uniprot dataframe.}
}
\description{
Given a character vector of Entrez IDs, this function will convert it to
Uniprot ID
}
\examples{
EG2Uniprot("9961")
}
\keyword{manip}

