% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensg2ensp.R
\name{ensg2ensp}
\alias{ensg2ensp}
\title{ensg2ensp}
\usage{
ensg2ensp(ensg, organism = "human")
}
\arguments{
\item{ensg}{ENSEMBL ID}

\item{organism}{supported organism}
}
\value{
data.frame
}
\description{
convert ENSEMBL to ENSEMBLROT
}
\author{
ygc
}

