% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getDesc.R
\name{getDesc}
\alias{getDesc}
\title{Query Description of genes}
\usage{
getDesc(id, type, organism = "human")
}
\arguments{
\item{id}{gene ids.}

\item{type}{id type.}

\item{organism}{supported organism.}
}
\value{
\item{desc}{a dataframe of gene id and corresponding description.}
}
\description{
Query Description of genes
}
\keyword{manip}

