% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getLeadingID}
\alias{getLeadingID}
\title{Get Leading ID in a ID strings.}
\usage{
getLeadingID(id, split = ";")
}
\arguments{
\item{id}{ID.}

\item{split}{split the IDs}
}
\value{
\item{lid}{Leading ID.}
}
\description{
Given a character string of several IDs, this function will split it and
return the first ID.
}
\examples{
getLeadingID("IPI00000684;IPI00607787;IPI00217816", split=";")
}
\seealso{
\code{\link{IPI2EG}}
}
\keyword{manip}

