% Generated by roxygen2 (4.0.2): do not edit by hand
\name{IDconverter}
\alias{IDconverter}
\title{IDconverter}
\usage{
IDconverter(ID, fromType, toType, organism)
}
\arguments{
\item{ID}{ID vector}

\item{fromType}{type of the input}

\item{toType}{type for converting}

\item{organism}{only human supported now}
}
\value{
data.frame
}
\description{
mapping ID
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

