% Generated by roxygen2 (4.0.2): do not edit by hand
\name{IPI2EG}
\alias{IPI2EG}
\title{Convert IPI to EG}
\usage{
IPI2EG(ipi, organism = "human", useBiomart = FALSE, withoutVersion = TRUE)
}
\arguments{
\item{ipi}{IPI.}

\item{organism}{organism}

\item{useBiomart}{using biomart or using annnotation package to convert ID}

\item{withoutVersion}{with or without IPI version}
}
\value{
\item{result}{IPI.Entrez dataframe}
}
\description{
Given a character vector of IPI, this function will convert it to Entrez
gene IDs.
}
\examples{
IPI2EG("IPI00470674")
}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\seealso{
\code{\link{Uniprot2EG}}
}
\keyword{manip}

