% Generated by roxygen2 (4.0.1): do not edit by hand
\name{removeVer}
\alias{removeVer}
\title{Remover Version}
\usage{
removeVer(id)
}
\arguments{
\item{id}{Gene ID.}
}
\value{
\item{id.nr}{Gene ID without version.}
}
\description{
Remover Version, such as IPI123123.1.
}
\examples{
removeVer("IPI00470674.2")
}
\keyword{manip}

