% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Uniprot2EG.R
\name{Uniprot2EG}
\alias{Uniprot2EG}
\title{Uniprot2EG}
\usage{
Uniprot2EG(uniprot, organism = "human", getDescription = FALSE,
  useBiomart = TRUE)
}
\arguments{
\item{uniprot}{uniprot id}

\item{organism}{only human supported now}

\item{getDescription}{get gene description or not}

\item{useBiomart}{using Biomart to query or using annotation packages.}
}
\value{
a data.frame for mapping
}
\description{
Convert uniprot ID to entrez gene id
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

