% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ACCNUM2EG}
\alias{ACCNUM2EG}
\title{ACCNUM2EG}
\usage{
ACCNUM2EG(ACCNUM, organism = "human")
}
\arguments{
\item{ACCNUM}{Accession numbers}

\item{organism}{organism, only human supported yet.}
}
\value{
a data.frame of accession to eg mapping
}
\description{
Convert ACCNUM to Entrez
}
\details{
Given a character vector of ACCNUM, this function will convert it to Entrez
gene IDs.
}
\examples{
ACCNUM2EG("AA086285")
}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\keyword{manip}

