% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IPI2Symbol.R
\name{IPI2Symbol}
\alias{IPI2Symbol}
\title{Convert IPI to Symbol_symbol}
\usage{
IPI2Symbol(ipi, organism = "human", useBiomart = FALSE,
  withoutVersion = TRUE)
}
\arguments{
\item{ipi}{IPI.}

\item{organism}{organism.}

\item{useBiomart}{using biomart or using annnotation package to convert ID}

\item{withoutVersion}{with or without IPI version}
}
\value{
\item{Symbol}{Symbol_symbol.}
}
\description{
Given a character vector of IPI, this function will convert it to
Symbol_symbol.
}
\examples{
IPI2Symbol("IPI00894498")
}
\seealso{
\code{\link{IPI2EG}}
}
\keyword{manip}

