% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Symbol2EG.R
\name{Symbol2EG}
\alias{Symbol2EG}
\title{Convert Symbol to Entrez}
\usage{
Symbol2EG(symbol, organism = "human")
}
\arguments{
\item{symbol}{Symbol.}

\item{organism}{organism.}
}
\value{
\item{eg}{Entrez IDs.}
}
\description{
Given a character vector of Symbol, this function will convert it to
Entrez IDs.
}
\examples{
Symbol2EG("GAPDH")
}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\seealso{
\code{\link{ACCNUM2EG}}
}
\keyword{manip}

